pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MIR"           # lowercase package name
version = "0.0.1"
description = "Medical-image registration toolkit"
authors = [{ name = "Junyu Chen", email = "jchen245@jhmi.edu" }]
readme = "README.md"
requires-python = ">=3.8.16"

# Dependencies
dependencies = [
	"pytest",
	"antspyx",
	"einops>=0.8.1",
	"matplotlib>=3.7.3",
	"ml_collections>=0.1.1",
	"nibabel",
	"numpy",
	"pymedio>=0.2.5",
	"scipy",
	"seaborn>=0.11.0",
	"SimpleITK>=2.2.1",
	"statsmodels>=0.14.0",
	"natsort",
	"torch>=2.0.1",
	"timm",
]

[project.optional-dependencies]
dev = ["black", "pytest"]

[tool.setuptools]
package-dir = {"" = "src"}    # <-- tell it that your packages live in src/

[tool.setuptools.package-data]
# this says: in the src/MIR/utils folder, include any .txt
"MIR.utils" = ["FreeSurfer_label_info.txt"]
"MIR.models" = ["deedsBCV/bin/*"]
"MIR.templates" = ["*.nii.gz"]

# either explicitly list your package
# packages = ["MIR"]
# or use find:
[tool.setuptools.packages.find]
where  = ["src"]             # <-- look under src/
include = ["MIR*",] 